requirements.txt
streamlit
langchain
langchain-community
langchain-openai
supabase
pypdf
python-dotenv
pyyaml